Fix ACP module import paths after refactoring
After moving acp/ into core/src/acp/, import paths in tests were incorrect:
- Fixed test_support imports: crate::test_support → crate::acp::test_support
- Fixed types imports in connection.rs tests: super::types → crate::acp::types
- Fixed ACP_MODE_EXTERNAL_AGENT references in detection.rs tests

All unit tests now pass (177 passed).

Also regenerated TypeScript bindings with ts-rs.
