+++ core/src/acp/config.rs
@@ -152,8 +152,8 @@ pub fn is_external_mode(mode: &str) -> bool {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::acp::test_support::ENV_LOCK;
     use crate::config::{APP_DIR_ENV, WORKSPACE_ROOT_ENV};
-    use crate::test_support::ENV_LOCK;
     use std::{env, fs};
 
     fn with_env_lock<T>(f: impl FnOnce() -> T) -> T {
